console: refactor Plugin tests
